rename the default module to standalone
