+++ src/redis/mod.rs
@@ -1,7 +1,7 @@
-mod default;
+mod standalone;
 mod stack;
 
 pub const REDIS_PORT: u16 = 6379;
 
-pub use default::Redis;
+pub use standalone::Redis;
 pub use stack::RedisStack;
